refactor: clear mechanical sonar smells

+++ batch-console-api/src/main/java/io/github/pinpols/batch/console/domain/rbac/support/LoginFailureTracker.java
@@ -96,7 +96,7 @@ public class LoginFailureTracker {
   }
 
   private long windowMillis() {
-    return (long) properties.getFailWindowMinutes() * 60_000L;
+    return properties.getFailWindowMinutes() * 60_000L;
   }
 
   private static String normalize(String raw) {
